# artifacts/openapi.yaml
openapi: 3.0.0
info:
  title: Nexus API
  version: "0.43"
  description: API for managing chat conversations, documents, and internet searches.
servers:
  - url: /api

paths:
  /api/chat:
    post:
      summary: Post a message to a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                message:
                  type: string
                conversation_id:
                  type: string
              required:
                - user_id
                - message
                - conversation_id
      responses:
        '200':
          description: Chat response
        '400':
          description: Missing user_id or message
        '500':
          description: Server error
          
  /api/chat/conversations:
    get:
      summary: Get conversations for a user
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of conversations
        '400':
          description: Missing user_id
        '404':
          description: No conversations found

  /api/chat/conversation/{conversation_id}:
    get:
      summary: Get conversation history by conversation_id
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation history
        '400':
          description: Missing user_id
        '404':
          description: Conversation not found or access denied

    delete:
      summary: Delete conversation by conversation_id
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation deleted successfully
        '400':
          description: Missing user_id
        '404':
          description: Conversation not found or access denied
        '500':
          description: Server error

  /api/chat/file:
    post:
      summary: Upload a file to a conversation
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                conversation_id:
                  type: string
                file:
                  type: string
                  format: binary
              required:
                - user_id
                - conversation_id
                - file
      responses:
        '200':
          description: File content added to conversation
        '400':
          description: Missing user_id, file, or conversation_id
        '403':
          description: Conversation not found or access denied
        '500':
          description: Server error

  /api/chat/internet:
    post:
      summary: Perform an internet search and add the results to a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                question:
                  type: string
                conversation_id:
                  type: string
              required:
                - user_id
                - question
                - conversation_id
      responses:
        '200':
          description: Internet search results added to conversation
        '400':
          description: Missing question, user_id, or conversation_id
        '404':
          description: No relevant information found
        '500':
          description: Server error

  /api/documents:
    get:
      summary: Get user documents
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of user documents
        '400':
          description: Missing user_id

    post:
      summary: Upload a new user document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                file:
                  type: string
                  format: binary
              required:
                - user_id
                - file
      responses:
        '200':
          description: Document uploaded
        '400':
          description: Missing user_id

  /api/documents/{document_id}:
    get:
      summary: Get a specific user document by document_id
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested document
        '400':
          description: Missing user_id
        '404':
          description: Document not found

    delete:
      summary: Delete a specific document by document_id
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document deleted successfully
        '400':
          description: Missing user_id
        '404':
          description: Document not found
        '500':
          description: Error deleting document

  /api/documents/{document_id}/versions:
    get:
      summary: Get all versions of a document by document_id
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of document versions
        '400':
          description: Missing user_id
        '404':
          description: No versions found
        '500':
          description: Error retrieving document versions

  /api/documents/{document_id}/version/{version}:
    get:
      summary: Get a specific version of a document
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested document version
        '400':
          description: Missing user_id
        '404':
          description: Document version not found
        '500':
          description: Error retrieving document version

    delete:
      summary: Delete a specific version of a document
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document version deleted successfully
        '400':
          description: Missing user_id
        '404':
          description: Document version not found
        '500':
          description: Error deleting document version

  /api/documents/{document_id}/chunks:
    get:
      summary: Get chunks of a document by document_id
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The chunks of the requested document
        '400':
          description: Missing user_id
        '404':
          description: No chunks found
        '500':
          description: Error retrieving chunks

  /api/documents/{document_id}/version/{version}/chunks:
    get:
      summary: Get chunks of a specific version of a document
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The chunks of the requested document version
        '400':
          description: Missing user_id
        '404':
          description: No chunks found for the specified document version
        '500':
          description: Error retrieving chunks

  /api/documents/search:
    post:
      summary: Search document chunks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                query:
                  type: string
                top_n:
                  type: integer
                  default: 5
              required:
                - user_id
                - query
      responses:
        '200':
          description: Search results
        '400':
          description: Missing user_id or query
        '500':
          description: Error during search